+++ app/src/main/java/com/hydrofish/app/animations/HelperDataClasses.kt
@@ -1,0 +1,9 @@
+package com.hydrofish.app.animations
+
+data class FishInfo(val coordinates: Coordinates, val fishId: Int)
+data class AnimationParams(
+    val xVal: Float = 0f,
+    val yVal: Float = 0f,
+    val rotateVal: Float = 0f,
+    val flipVal: Float = 0f
+)
